Refactor #293 rename Generated method
