+++ app/src/main/java/com/nlab/reminder/domain/feature/home/HomeWorkflow.kt
@@ -16,7 +16,7 @@
 
 package com.nlab.reminder.domain.feature.home
 
-import com.nlab.reminder.core.annotation.test.ExcludeGeneratedFromTestReport
+import com.nlab.reminder.core.kotlin.annotation.Generated
 import com.nlab.reminder.core.data.model.Tag
 
 /**
@@ -28,13 +28,13 @@ internal sealed interface HomeWorkflow {
     data object TimetableSchedule : HomeWorkflow
     data object AllSchedule : HomeWorkflow
 
-    @ExcludeGeneratedFromTestReport
+    @Generated
     data class TagConfig(
         val tag: Tag,
         val usageCount: Long
     ) : HomeWorkflow
 
-    @ExcludeGeneratedFromTestReport
+    @Generated
     data class TagRename(
         val tag: Tag,
         val usageCount: Long,
@@ -42,7 +42,7 @@ internal sealed interface HomeWorkflow {
         val shouldKeyboardShown: Boolean
     ) : HomeWorkflow
 
-    @ExcludeGeneratedFromTestReport
+    @Generated
     data class TagDelete(
         val tag: Tag,
         val usageCount: Long
