+++ app/src/main/java/com/nlab/reminder/domain/feature/schedule/all/AllScheduleUiState.kt
@@ -16,7 +16,7 @@
 
 package com.nlab.reminder.domain.feature.schedule.all
 
-import com.nlab.reminder.core.annotation.test.ExcludeGeneratedFromTestReport
+import com.nlab.reminder.core.kotlin.annotation.Generated
 import com.nlab.reminder.core.annotation.platform.Stable
 import com.nlab.reminder.core.schedule.model.ScheduleElement
 import com.nlab.statekit.State
@@ -29,7 +29,7 @@ import kotlinx.collections.immutable.ImmutableList
 sealed interface AllScheduleUiState : State {
     object Empty : AllScheduleUiState
 
-    @ExcludeGeneratedFromTestReport
+    @Generated
     data class Loaded(
         val scheduleElements: ImmutableList<ScheduleElement>,
         val isCompletedScheduleShown: Boolean,
