+++ app/src/main/java/com/nlab/reminder/domain/feature/schedule/all/AllScheduleWorkflow.kt
@@ -16,13 +16,13 @@
 
 package com.nlab.reminder.domain.feature.schedule.all
 
-import com.nlab.reminder.core.annotation.test.ExcludeGeneratedFromTestReport
+import com.nlab.reminder.core.kotlin.annotation.Generated
 import com.nlab.reminder.core.data.model.Link
 
 /**
  * @author thalys
  */
 sealed interface AllScheduleWorkflow {
-    @ExcludeGeneratedFromTestReport
+    @Generated
     data class LinkPage(val link: Link) : AllScheduleWorkflow
 }
