+++ core/data/src/main/kotlin/com/nlab/reminder/core/data/model/LinkMetadataTable.kt
@@ -16,11 +16,11 @@
 
 package com.nlab.reminder.core.data.model
 
-import com.nlab.reminder.core.annotation.test.ExcludeGeneratedFromTestReport
+import com.nlab.reminder.core.kotlin.annotation.Generated
 
 /**
  * @author Doohyun
  */
-@ExcludeGeneratedFromTestReport
+@Generated
 @JvmInline
 value class LinkMetadataTable(val value: Map<Link, LinkMetadata>)
