+++ core/data/src/main/kotlin/com/nlab/reminder/core/data/model/Schedule.kt
@@ -16,13 +16,13 @@
 
 package com.nlab.reminder.core.data.model
 
-import com.nlab.reminder.core.annotation.test.ExcludeGeneratedFromTestReport
+import com.nlab.reminder.core.kotlin.annotation.Generated
 import kotlinx.collections.immutable.ImmutableList
 
 /**
  * @author Doohyun
  */
-@ExcludeGeneratedFromTestReport
+@Generated
 data class Schedule(
     val id: ScheduleId,
     val title: String,
